implementacion de softDElete de tabla expenses o gastos, implemtacion de carousel con productos de categorya comida rapida, para detallar info de compras

+++ app/Http/Controllers/RecordsController.php
@@ -35,19 +35,26 @@ class RecordsController extends Controller
       $expense->account_id=$request->user()->account->id;
       $expense->quantity=$request->quantity;
       if ($expense->save()) {
-        return redirect ('/record/info/'. $request->subcategory_id)->with(['message'=>'Se agregagó el registro correctamente', 'alert'=>'warning']);
+        return redirect ('/record/info/'. $request->subcategory_id)->with(['message'=>'Se agregagó el registro correctamente', 'alert'=>'warning', 'expense'=>$expense]);
       }
       else {
         return redirect ('/cat/')->with(['message'=>'Ocurrió un error al guardar el registro', 'alert'=>'danger']);
       }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request )
     {
         //
-        $expense= Expense::find(1);
-        if ($expense->delete()) {
-          dd('borrado');
+        $expense= Expense::find($request->id);
+        if ($expense) {
+          if ($expense->delete()) {
+            return redirect ('/'. $request->subcategory_id)->with(['message'=>'Se eliminó un registro', 'alert'=>'danger']);
+          }
+          else {
+            return redirect ('/')->with(['message'=>'Ocurrió un error al eliminar el registro', 'alert'=>'danger']);
+          }
+        }else {
+          return redirect ('/')->with(['message'=>'Algo salió mal, recargue la página por favor', 'alert'=>'danger']);
         }
     }
 }
